pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stitchtoon"
version = "2.1.0"
authors = [
  { name="Beshr Alghalil", email="beshrghalil@protonmail.com" },
]
description = "A powerful program for stitching and cutting webtoons/manhwa/manhua raws."
readme = "README.md"
requires-python = ">=3.8"
keywords = ["manga", "webtoon", "stitch", "slice", "combin"]
dependencies = [
    "psd_tools",
    'natsort',
    'pillow',
    'progress',
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/BishrGhalil/stitchtoon"
"Bug Tracker" = "https://github.com/BishrGhalil/issues"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "black",
    "flake8",
    "isort",
    "tox",
    "mypy",
]

[project.scripts]
stitchtoon = "stitchtoon.__main__:main"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.pylint.format]
max-line-length = "120"
